requirements.txt
flask
audio-separator==0.29.0
pydub
noisereduce
scipy
numpy<2.0
librosa
werkzeug
pywebview
pyinstaller
psutil
mutagen
faster-whisper
soundfile
